+++ packages/terra-button-group/CHANGELOG.md
@@ -3,6 +3,17 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Supports single-select and mulit-select.
+* `non-selectable`,  `single-select`, `multi-select` select types in ButtonGroup
+* isDisabled, onBlur, onFocus, onKeyDown, onKeyUp props in ButtonGroup.Button
+
+### Removed
+* buttons, isCompact, size, variant props in ButtonGroup
+* children, isReversed props in ButtonGroup.Button
+
+### Changed
+* text prop is now required.
 
 1.16.0 - (January 5, 2018)
 ------------------